backend/requirements.txt
# core framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
gunicorn==21.2.0
pydantic==2.6.1
pydantic-settings==2.1.0

# database & ORM
prisma==0.13.1
sqlalchemy==2.0.27
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AI/ML stack
ultralytics==8.1.20  # YOLOv8
google-generativeai>=0.8.0  # Gemini API
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
numpy==1.26.4
pillow==10.2.0
scikit-learn==1.4.0
pandas==2.2.0
torch==2.2.0
torchvision==0.17.0

# redis & caching
redis==5.0.1
hiredis==2.3.2
aioredis==2.0.1

# celery & background tasks
celery==5.3.6
flower==2.0.1  # celery monitoring

# websocket & real-time
python-socketio==5.11.1
fastapi-socketio==0.0.10
websockets==12.0

# authentication & security
python-jose[cryptography]==3.3.0
passlib[bcrypt]>=1.7.4
python-multipart==0.0.9
pyjwt==2.8.0
bcrypt>=4.2.0

# HTTP client
httpx==0.26.0
aiohttp==3.9.3

# notifications
twilio==8.12.0
sendgrid==6.11.0
firebase-admin==6.4.0

# file processing
python-magic==0.4.27
ffmpeg-python==0.2.0

# monitoring & logging
sentry-sdk[fastapi]==1.40.0
prometheus-client==0.19.0
structlog==24.1.0
python-json-logger==2.0.7

# utilities
python-dotenv==1.0.1
email-validator==2.1.0
pytz==2024.1
requests==2.31.0

# testing
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
# httpx-test==0.1.0
faker==22.6.0